requirement.txt
pyyaml
dependency_injector